+++ CHANGELOG.md
@@ -8,6 +8,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## Unreleased
 
+## 0.17.0 - 2025-06-17
+
 ***Added:***
 
 - Persist the VS Code extensions directory in the cache for the `linux-container` developer environment type